Adds player_factory, functionality moved from app.rb
